#7 update web3j wrappers
 - Remove `MyTransientClient` class from noosphere agent contracts
